move regex impl to it own package
