+++ app/src/main/java/com/realityexpander/tasky/auth_feature/domain/validation/validateEmail/ValidateEmailImpl.kt
@@ -1,7 +1,6 @@
 package com.realityexpander.tasky.auth_feature.domain.validation.validateEmail
 
 import com.realityexpander.tasky.auth_feature.domain.validation.validateEmail.emailMatcherImpls.EmailMatcherAndroidImpl
-import com.realityexpander.tasky.auth_feature.domain.validation.validateEmail.emailMatcherImpls.EmailMatcherRegexImpl
 
 class ValidateEmailImpl (
     private val emailMatcher: IEmailMatcher = EmailMatcherAndroidImpl()
@@ -9,14 +8,5 @@ class ValidateEmailImpl (
     override fun validate(email: String): Boolean {
         return email.isNotBlank()
             && emailMatcher.matches(email)
-    }
-}
-
-class ValidateEmailRegexImpl (
-    private val emailMatcher: IEmailMatcher = EmailMatcherRegexImpl()
-): IValidateEmail {
-    override fun validate(email: String): Boolean {
-        return email.isNotBlank()
-                && emailMatcher.matches(email)
     }
 }
