+++ app/src/main/java/com/realityexpander/tasky/auth_feature/domain/validation/validateEmail/ValidateEmailRegexImpl.kt
@@ -1,0 +1,12 @@
+package com.realityexpander.tasky.auth_feature.domain.validation.validateEmail
+
+import com.realityexpander.tasky.auth_feature.domain.validation.validateEmail.emailMatcherImpls.EmailMatcherRegexImpl
+
+class ValidateEmailRegexImpl (
+    private val emailMatcher: IEmailMatcher = EmailMatcherRegexImpl()
+): IValidateEmail {
+    override fun validate(email: String): Boolean {
+        return email.isNotBlank()
+                && emailMatcher.matches(email)
+    }
+}
